+++ conf/rad/specs/1.0.1.spec
@@ -37,10 +37,12 @@ if( [ $RPM_BUILD_ROOT != '/' ] ); then rm -rf $RPM_BUILD_ROOT; fi;
 if [ "$1" = "1" ]; then
   # Perform tasks to prepare for the initial installation
   echo "Installing rad user environment..."
-  useradd -g apache -M -s /bin/false rad
+  if [ "$(grep "rad" /etc/passwd | wc -l)" -eq 0 ]; then
+  	useradd -g apache -M -r -s /bin/false rad
+  fi
   
   # Create the log folder for rad
-  if [ ! -d "$DIRECTORY" ]; then
+  if [ ! -d "/var/log/rad" ]; then
 	  mkdir /var/log/rad
 	  chown rad:apache /var/log/rad
 	  chmod 775 /var/log/rad
@@ -55,14 +57,12 @@ if [ "$1" = "1" ]; then
   # Copy the logrotate.d so that we can rotate our logs
   cp -f /home/rad/admin/init/config/logrotate /etc/logrotate.d/rad
 
-  # Copy the crontab so that we can start our processes
-  cp -f /home/rad/admin/init/config/crontab /etc/cron.d/rad
 elif [ "$1" = "2" ]; then
   # Perform whatever maintenance must occur before the upgrade begins
   echo "Upgrading rad user environment..."
   
   # Create the log folder for rad
-  if [ ! -d "$DIRECTORY" ]; then
+  if [ ! -d "/var/log/rad" ]; then
 	  mkdir /var/log/rad
 	  chown rad:apache /var/log/rad
 	  chmod 775 /var/log/rad
@@ -75,10 +75,7 @@ elif [ "$1" = "2" ]; then
   # Copy the logrotate.d so that we can rotate our logs
   rm -f /etc/logrotate.d/rad
   cp -f /home/rad/admin/init/config/logrotate /etc/logrotate.d/rad
-  
-  # Copy the crontab so that we can start our processes
-  rm -f /etc/cron.d/rad
-  cp -f /home/rad/admin/init/config/crontab /etc/cron.d/rad
+
 fi
 
 
@@ -86,10 +83,11 @@ fi
 if [ "$1" = "0" ]; then
   # Perform tasks to prepare for the final uninstallation
   echo "Removing rad user environment..."
-  rm -f /etc/cron.d/rad
   rm -f /etc/logrotate.d/rad
   rm -Rf /var/log/rad
-  userdel -r rad
+  if [ "$(grep "rad" /etc/passwd | wc -l)" -gt 0 ]; then
+    userdel -r rad
+  fi
 elif [ "$1" = "2" ]; then
   # Perform whatever maintenance must occur before the upgrade begins
   echo "Upgrading rad user environment..."
